Interpolation search

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -21,7 +21,7 @@ int interpolation_search(int *array, size_t size, int value)
 	{
 		pos = low + (((double)(high - low) / (array[high] - array[low])) *
 			    (value - array[low]));
-		printf("Value checked array[%lu] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long)pos, array[pos]);
 
 		if (array[pos] == value)
 			return (pos);
@@ -31,6 +31,6 @@ int interpolation_search(int *array, size_t size, int value)
 			high = pos - 1;
 	}
 
-	printf("Value checked array[%lu] is out of range\n", pos);
+	printf("Value checked array[%lu] is out of range\n", (unsigned long)pos);
 	return (-1);
 }
